Improve refreshAuth response workflow and minor changes in utils
