+++ utils/fetch_remote.go
@@ -119,26 +119,3 @@ func runCmd(command string, args []string) (string, error) {
 	return strings.TrimRight(string(output), "\000"), nil
 
 }
-
-// Utility to parse the --repo flag
-func RepoArgumentResolver(arg string) ([]string, error) {
-
-	// github.com/deepsourcelabs/cli or gh/deepsourcelabs/cli
-
-	argComponents := strings.Split(arg, "/")
-
-	switch argComponents[0] {
-	case "gh", "github.com":
-		argComponents[0] = "GITHUB"
-
-	case "gl", "gitlab.com":
-		argComponents[0] = "GITLAB"
-
-	case "bb", "bitbucket.com":
-		argComponents[0] = "BITBUCKET"
-	default:
-		return argComponents, fmt.Errorf("VCSProvider `%s` not supported", argComponents[0])
-	}
-
-	return argComponents, nil
-}
