+++ utils/remote_resolver.go
@@ -69,3 +69,26 @@ func ResolveRemote(repoArg string) (*RemoteData, error) {
 	}
 	return &remote, nil
 }
+
+// Utility to parse the --repo flag
+func RepoArgumentResolver(arg string) ([]string, error) {
+
+	// github.com/deepsourcelabs/cli or gh/deepsourcelabs/cli
+
+	argComponents := strings.Split(arg, "/")
+
+	switch argComponents[0] {
+	case "gh", "github.com":
+		argComponents[0] = "GITHUB"
+
+	case "gl", "gitlab.com":
+		argComponents[0] = "GITLAB"
+
+	case "bb", "bitbucket.com":
+		argComponents[0] = "BITBUCKET"
+	default:
+		return argComponents, fmt.Errorf("VCSProvider `%s` not supported", argComponents[0])
+	}
+
+	return argComponents, nil
+}
